+++ inst/system/installSuggestedPackages.R
@@ -10,5 +10,5 @@ install.packages("nnls")
 install.packages("e1071")
 install.packages("polspline")
 install.packages("quadprog")
-install.packages("randomForest")
+install.packages("ranger")